+++ medical-record/medical-record/src/main/java/com/ivanovp/medical_record/dto/request/ChangeGpDTO.java
@@ -1,0 +1,15 @@
+package com.ivanovp.medical_record.dto.request;
+
+import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class ChangeGpDTO {
+
+    @NotNull
+    private Long gpId;
+}