+++ medical-record/medical-record/src/main/java/com/ivanovp/medical_record/dto/request/PatientCreateDTO.java
@@ -1,0 +1,33 @@
+package com.ivanovp.medical_record.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Pattern;
+import jakarta.validation.constraints.Size;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class PatientCreateDTO {
+
+    @NotBlank
+    private String username;
+
+    @NotBlank
+    @Size(min = 6)
+    private String password;
+
+    @NotBlank
+    private String name;
+
+    @NotBlank
+    @Pattern(regexp = "\\d{10}")
+    private String egn;
+
+    private boolean isInsured;
+
+    private Long gpId;
+}