+++ medical-record/medical-record/src/main/java/com/ivanovp/medical_record/dto/response/ExaminationResponseDTO.java
@@ -1,0 +1,13 @@
+package com.ivanovp.medical_record.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class ExaminationResponseDTO {
+
+    private Long id;
+}